Restructured API set-up for events
